+++ packages/block-editor/src/components/block-card/index.js
@@ -3,16 +3,14 @@
  */
 import BlockIcon from '../block-icon';
 
-function BlockCard( { blockType } ) {
+function BlockCard( { blockType: { icon, title, description } } ) {
 	return (
 		<div className="block-editor-block-card">
-			<BlockIcon icon={ blockType.icon } showColors />
+			<BlockIcon icon={ icon } showColors />
 			<div className="block-editor-block-card__content">
-				<h2 className="block-editor-block-card__title">
-					{ blockType.title }
-				</h2>
+				<h2 className="block-editor-block-card__title">{ title }</h2>
 				<span className="block-editor-block-card__description">
-					{ blockType.description }
+					{ description }
 				</span>
 			</div>
 		</div>