+++ packages/edit-site/src/components/navigation-sidebar/navigation-panel/templates-navigation.js
@@ -1,13 +1,13 @@
 /**
  * WordPress dependencies
  */
-import { useRef, useEffect } from '@wordpress/element';
 import {
 	__experimentalNavigation as Navigation,
 	__experimentalNavigationMenu as NavigationMenu,
 	__experimentalNavigationItem as NavigationItem,
 	__experimentalNavigationBackButton as NavigationBackButton,
 } from '@wordpress/components';
+import { __experimentalMainDashboardButton as MainDashboardButton } from '@wordpress/interface';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { __ } from '@wordpress/i18n';
 
@@ -19,14 +19,6 @@ import TemplatePartsMenu from './menus/template-parts';
 import { MENU_ROOT, MENU_TEMPLATE_PARTS, MENU_TEMPLATES } from './constants';
 
 export default function TemplatesNavigation() {
-	const ref = useRef();
-
-	useEffect( () => {
-		if ( ref.current ) {
-			ref.current.focus();
-		}
-	}, [ ref ] );
-
 	const { templateId, templatePartId, templateType, activeMenu } = useSelect(
 		( select ) => {
 			const {
@@ -59,12 +51,13 @@ export default function TemplatesNavigation() {
 			onActivateMenu={ setNavigationPanelActiveMenu }
 		>
 			{ activeMenu === MENU_ROOT && (
-				<NavigationBackButton
-					backButtonLabel={ __( 'Dashboard' ) }
-					className="edit-site-navigation-panel__back-to-dashboard"
-					href="index.php"
-					ref={ ref }
-				/>
+				<MainDashboardButton.Slot>
+					<NavigationBackButton
+						backButtonLabel={ __( 'Dashboard' ) }
+						className="edit-site-navigation-panel__back-to-dashboard"
+						href="index.php"
+					/>
+				</MainDashboardButton.Slot>
 			) }
 
 			<NavigationMenu title={ __( 'Theme' ) }>