+++ packages/react-native-bridge/ios/RNReactNativeGutenbergBridge.swift
@@ -289,6 +289,43 @@ public class RNReactNativeGutenbergBridge: RCTEventEmitter {
     }
 
     @objc
+    func requestMediaFilesEditorLoad(_ mediaFiles: [String], blockId: String) {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestMediaFilesEditorLoad(mediaFiles, blockId: blockId)
+        }
+    }
+
+    @objc
+    func requestMediaFilesFailedRetryDialog(_ mediaFiles: [String]) {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestMediaFilesFailedRetryDialog(mediaFiles)
+        }
+    }
+
+    @objc
+    func requestMediaFilesUploadCancelDialog(_ mediaFiles: [String]) {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestMediaFilesUploadCancelDialog(mediaFiles)
+        }
+    }
+
+    @objc
+    func requestMediaFilesSaveCancelDialog(_ mediaFiles: [String]) {
+        DispatchQueue.main.async {
+            self.delegate?.gutenbergDidRequestMediaFilesSaveCancelDialog(mediaFiles)
+        }
+    }
+
+    @objc
+    func mediaSaveSync() {
+        DispatchQueue.main.async {
+            if self.hasObservers {
+                self.delegate?.gutenbergDidRequestMediaSaveSync()
+            }
+        }
+    }
+
+    @objc
     func actionButtonPressed(_ buttonType: String) {
         guard let button = Gutenberg.ActionButtonType(rawValue: buttonType) else {
             return
@@ -321,6 +358,7 @@ extension RNReactNativeGutenbergBridge {
         case replaceBlock
         case updateCapabilities
         case showNotice
+        case mediaSave
     }
 
     public override func supportedEvents() -> [String]! {
@@ -361,23 +399,17 @@ extension RNReactNativeGutenbergBridge {
     }
 }
 
-extension RNReactNativeGutenbergBridge {
-    enum MediaKey {
-        static let id = "id"
-        static let url = "url"
-        static let type = "type"
-        static let caption = "caption"
-    }
-}
-
 extension MediaInfo {
-
+    /// Dynamically wraps up all properties into a Json Object to be sent to JS Side.
     func encodeForJS() -> [String: Any] {
-        return [
-            RNReactNativeGutenbergBridge.MediaKey.id: id as Any,
-            RNReactNativeGutenbergBridge.MediaKey.url: url as Any,
-            RNReactNativeGutenbergBridge.MediaKey.type: type as Any,
-            RNReactNativeGutenbergBridge.MediaKey.caption: caption as Any
-        ]
+        guard
+            let data = try? JSONEncoder().encode(self),
+            let jsonObject = try? JSONSerialization.jsonObject(with: data, options: []) as? [String: Any] else
+        {
+            assertionFailure("Encoding of MediaInfo failed")
+            return [String: Any]()
+        }
+
+        return jsonObject
     }
 }